requirements.txt
python-dotenv>=1.0.0
livekit-api>=0.3.0
livekit-plugins>=0.1.0
livekit-agents>=0.1.0 